Move inner classes and
